+++ src/main/java/de/maxhenkel/audioplayer/AudioPlayerPermissionManager.java
@@ -16,11 +16,13 @@ public class AudioPlayerPermissionManager implements PermissionManager<CommandSo
 
     private static final Permission UPLOAD_PERMISSION = new Permission("audioplayer.upload", PermissionType.EVERYONE);
     private static final Permission APPLY_PERMISSION = new Permission("audioplayer.apply", PermissionType.EVERYONE);
+    private static final Permission APPLY_ANNOUNCER_PERMISSION = new AnnouncerPermission("audioplayer.apply_announcer", PermissionType.EVERYONE);
     private static final Permission PLAY_COMMAND_PERMISSION = new Permission("audioplayer.play_command", PermissionType.OPS);
 
     private static final List<Permission> PERMISSIONS = List.of(
             UPLOAD_PERMISSION,
             APPLY_PERMISSION,
+            APPLY_ANNOUNCER_PERMISSION,
             PLAY_COMMAND_PERMISSION
     );
 
@@ -76,6 +78,18 @@ public class AudioPlayerPermissionManager implements PermissionManager<CommandSo
                 default:
                     return false;
             }
+        }
+    }
+
+    private static class AnnouncerPermission extends Permission {
+
+        public AnnouncerPermission(String permission, PermissionType type) {
+            super(permission, type);
+        }
+
+        @Override
+        public boolean hasPermission(@Nullable ServerPlayer player) {
+            return AudioPlayer.SERVER_CONFIG.announcerDiscsEnabled.get() && super.hasPermission(player);
         }
     }